% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_distance.R
\name{jaccard_dist}
\alias{jaccard_dist}
\title{Compute Jaccard distance between two gene sets}
\usage{
jaccard_dist(A, B)
}
\arguments{
\item{A}{Character vector of genes.}

\item{B}{Character vector of genes.}
}
\value{
Numeric scalar distance between 0 and 1. A higher distance value indicates a bigger difference.
}
\description{
Compute Jaccard distance between two gene sets
}
